Protect against missing peaks
Low-quality samples for which peaks could not be identified caused
flowPloidy to choke when importing. The changes in this commit provide some
protection against this - allowing an empty peaks table to pass through the
initialization process. I also provide code to deal with missing peaks in
plotting functions. There are probably other problems left to ferret out
and correct!