Switched numericInput for selectInput for Sample Number
Tracey found a way to break the numericInput - the textbox is editable,
allowing the user to enter invalid values. Typically, she would select the
textbox, and delete the current value before entering the new one. As soon
as the old value was deleted, shiny would update the value to NULL, not
waiting for the new value to be entered. This caused the app to crash
completely and irrecoverably. Switching to selectInput, the user can only
pick between the valid options, they don't have an opportunity to enter
arbitrary values (or NULL), so they can't break things (as easily).