Added new fhDoCounts2 function
Iterates over all components and calculates cell counts for those that have
the `doCounts` flag set to TRUE. No longer requires that I explicitly add
components that I want counted to a list inside the function.

Still to do - fix up the places for fhCounts is accessed: exFlowHist, and
the show method for FlowHist objects.