Added support for truncate_max_range and trimRaw
`truncate_max_range` is an argument for flowCore::read.FCS. Some fcs files
include raw data for a channel that exceeds the stated limits in the header
meta data. By default, `read.FCS` rejects such files with an error. With
`truncate_max_range == FALSE`, it will allow reading such files.

This commit adds `truncate_max_range` as an option for `FlowHist`, which
passes the value of the argument directly to `read.FCS`.

Some files that exceed their stated maximum values appear to include a lot
of empty bins. I've added the argument/slot `trimRaw` to `FlowHist`
arguments, which lets the user set a threshold for excluding such data.
There should be some way to determine the appropriate value for this from
the instruments documentation?