Improved peak detection for endopolyploidy
The original peak finding algorithm explicitly excluded peaks that were
double/half the size of other peaks. This improves model fitting when
working with G1/G2 peaks. However, it makes peak detection useless for
endopolyploidy, where the peaks are all multiples of each other! So I added
a check in cleanPeaks that skips the pruning of double/half size peaks.